+++ Src/TimberApi.Tools/ToolSystem/Tools/GenericTool/GenericToolConfigurator.cs
@@ -1,0 +1,14 @@
+using Bindito.Core;
+using TimberApi.SpecificationSystem;
+using TimberApi.Tools.ToolSystem.Tools.Planting;
+
+namespace TimberApi.Tools.ToolSystem.Tools.GenericTool;
+
+[Context("Game")]
+public class GenericToolConfigurator : IConfigurator
+{
+    public void Configure(IContainerDefinition containerDefinition)
+    {
+        containerDefinition.MultiBind<IToolFactory>().To<GenericToolFactory>().AsSingleton();
+    }
+}