+++ Src/TimberApi.Tools/ToolSystem/Tools/GenericTool/GenericToolToolInformation.cs
@@ -1,0 +1,6 @@
+namespace TimberApi.Tools.ToolSystem.Tools.GenericTool;
+
+public class GenericToolToolInformation(string className)
+{
+    public string ClassName { get; } = className;
+}