+++ Src/TimberApi.Tools/ToolSystem/Tools/SettingBox/SettingBoxTool.cs
@@ -3,20 +3,11 @@ using Timberborn.ToolSystem;
 
 namespace TimberApi.Tools.ToolSystem.Tools.SettingBox;
 
-public class SettingBoxTool : Tool, IUnselectableTool
+public class SettingBoxTool(IOptionsBox optionsBox) : Tool, IUnselectableTool
 {
-    private readonly IOptionsBox _optionsBox;
-
-
-    public SettingBoxTool(ToolGroup? toolGroup, IOptionsBox optionsBox)
-    {
-        ToolGroup = toolGroup;
-        _optionsBox = optionsBox;
-    }
-
     public override void Enter()
     {
-        _optionsBox.Show();
+        optionsBox.Show();
     }
 
     public override void Exit()